Refactor file names, move features out of test directory
