+++ src/components/features/featuredetails/featuredetails.css
