+++ src/components/features/viewfeatures/Features.js
@@ -7,18 +7,18 @@ import './features.css';
 
 import {compose} from "redux";
 
-import createItemIcon from "../../../../assets/Icons/create.png";
+import createItemIcon from "../../../assets/Icons/create.png";
 import {
   showErrorAlert,
   showSuccessAlert
-} from "../../../../store/actions/snackbarActions";
-import {setPrevUrl} from "../../../../store/actions/authActions";
-import Feature from "../../Feature";
-import {getServiceNameFromPathName} from "../../../../util/StringUtil";
+} from "../../../store/actions/snackbarActions";
+import {setPrevUrl} from "../../../store/actions/authActions";
+import Feature from "../Feature";
+import {getServiceNameFromPathName} from "../../../util/StringUtil";
 import {
   getFeaturesByService, resetGetFeaturesByService,
   unsubscribeGetFeaturesByService
-} from "../../../../store/actions/featureActions";
+} from "../../../store/actions/featureActions";
 
 const Features = (props) => { //todo move this out of tests parent package/dir
